split prmitives to seperate files & remove lodash
