+++ packages/watch/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@solid-primitives/watch",
   "version": "0.0.1",
-  "description": "Reactive helpers built around watching signal changes.",
+  "description": "Reactive helpers built around watching computation changes.",
   "author": "David Di Biase <dave.dibiase@gmail.com>",
   "license": "MIT",
   "homepage": "https://github.com/davedbase/solid-primitives/tree/main/packages/watch",
@@ -26,7 +26,6 @@
     "primitives"
   ],
   "devDependencies": {
-    "@types/lodash": "^4.14.176",
     "jsdom": "^18.0.1",
     "prettier": "^2.0.5",
     "solid-register": "^0.0.10",
@@ -38,7 +37,7 @@
     "solid-js": "^1.2.3"
   },
   "dependencies": {
-    "lodash.debounce": "^4.0.8",
-    "lodash.throttle": "^4.1.1"
+    "@solid-primitives/debounce": "^1.0.2",
+    "@solid-primitives/throttle": "^1.0.3"
   }
 }
