+++ packages/watch/src/watch.ts
@@ -1,0 +1,30 @@
+import { createEffect, createRoot, on, onCleanup, ReturnTypes } from "solid-js";
+import {
+  StopWatch,
+  ValidWatchCallback,
+  WatchArrayCallback,
+  WatchOptions,
+  WatchSignalCallback
+} from "./common";
+
+export function watch<T extends (() => any)[], U>(
+  source: [...T],
+  fn: WatchArrayCallback<T, U>,
+  options?: WatchOptions
+): StopWatch;
+
+export function watch<T extends () => any, U>(
+  source: T,
+  fn: WatchSignalCallback<T, U>,
+  options?: WatchOptions
+): StopWatch;
+
+export function watch(source: any, fn: ValidWatchCallback, options: WatchOptions = {}): StopWatch {
+  const stop = createRoot(stop => {
+    const { defer = true } = options;
+    createEffect(on(source, fn, { defer }), options);
+    return stop;
+  });
+  onCleanup(stop);
+  return stop;
+}
